QR Code  Fraud detection PROJECT/requirements.txt
# Streamlit and basic dependencies
streamlit
numpy
pandas

# Computer Vision for QR Code Analysis
opencv-python
Pillow
pyzbar

# QR Code Generation and Reading
qrcode
qrcode[pil]

# URL Analysis (no additional dependencies needed for url_analysis.py)

# QR Code API Integration (for free QR code decoding APIs)
requests
